# Conflicts:
#	VerticalWarehouses/Panel PC UI/InstallationApp/Resources/InstallationAppModule.cs
#	VerticalWarehouses/Panel PC UI/InstallationApp/ViewsAndViewModels/ShuttersControl/Shutter1ControlViewModel.cs

+++ WMS/Data/Core/Interfaces/ICompartmentProvider.cs
@@ -25,7 +25,7 @@ namespace Ferretto.WMS.Data.Core.Interfaces
 
         Task<IEnumerable<CompartmentDetails>> GetByLoadingUnitIdAsync(int id);
 
-        Task<int?> GetMaxCapacityAsync(double width, double height, int itemId);
+        Task<double?> GetMaxCapacityAsync(double width, double height, int itemId);
 
         #endregion
     }
