+++ WMS/Data/Core/Models/Item/ItemCompartmentType.cs
@@ -8,7 +8,7 @@ namespace Ferretto.WMS.Data.Core.Models
 
         public int ItemId { get; set; }
 
-        public int? MaxCapacity { get; set; }
+        public double? MaxCapacity { get; set; }
 
         #endregion
     }
